+++ libs/angular/src/v-angular/account-number/account-number.module.ts
@@ -1,0 +1,11 @@
+import { CommonModule } from '@angular/common'
+import { NgModule } from '@angular/core'
+
+import { AccountNumberPipe } from './account-number.pipe'
+
+@NgModule({
+  declarations: [],
+  imports: [CommonModule, AccountNumberPipe],
+  exports: [AccountNumberPipe],
+})
+export class NggvAccountNumberPipeModule {}