+++ libs/core/.storybook/manager.ts
@@ -28,10 +28,12 @@ const betaItems = {
   Image: true,
   Mask: true,
   Signal: true,
+  Select: true,
   Spacer: true,
   Text: true,
   Theme: true,
   Video: true,
+  Div: true,
 }
 
 addons.setConfig({