+++ libs/core/src/components/dropdown/dropdown.stories.ts
@@ -2,6 +2,8 @@ import { html } from 'lit'
 
 import type { Meta, StoryObj } from '@storybook/web-components'
 
+import { argTablePropsFor } from '../../../.storybook/argTableProps.ts'
+
 import './index.ts'
 import '../icon/icons/push.ts'
 
@@ -24,6 +26,9 @@ const meta: Meta = {
     layout: 'centered',
   },
   tags: ['autodocs'],
+  argTypes: {
+    ...argTablePropsFor('gds-dropdown'),
+  },
 }
 
 export default meta
@@ -264,6 +269,56 @@ export const Multiple: Story = {
       <gds-option value="cloning">Cloning</gds-option>
       <gds-option value="cryonics">Cryonics</gds-option>
       <gds-option value="teleportation">Teleportation</gds-option>
+    </gds-dropdown>
+  `,
+}
+
+/**
+ * The dropdown supports the combobox pattern with the `combobox` attribute.
+ *
+ * Use combobox when the user needs to have the option of entering a custom value but is also helped by a list of predefined options.
+ *
+ * Don't use combobox if you only need a searchable dropdown but don't want to allow custom values. Use the `searchable` attribute instead in that case.
+ *
+ * ***Note:***
+ * - The `combobox` attribute doesn't work with the `multiple` attribute. If both are used, the `combobox` attribute will be ignored.
+ * - Custom trigger content will not be rendered when the `combobox` attribute is used, since it will be replaced with an input element internally.
+ * - Never combine the `searchable` attribute with the `combobox` attribute, since the combobox input is already used for filtering.
+ * - For practical reasons, mobile styles are disabled when the `combobox` attribute is used.
+ */
+export const Combobox: Story = {
+  ...DefaultParams,
+  render: (args) => html`
+    <gds-dropdown label="Favorite sci-fi tech" combobox>
+      <gds-option value="" isplaceholder></gds-option>
+      <gds-option value="warp">Warp Drive</gds-option>
+      <gds-option value="cybernetics">Cybernetics</gds-option>
+      <gds-option value="nanotechnology">Nanotechnology</gds-option>
+      <gds-option value="cloning">Cloning</gds-option>
+      <gds-option value="cryonics">Cryonics</gds-option>
+      <gds-option value="teleportation">Teleportation</gds-option>
+      <gds-option value="ai">Artificial Intelligence</gds-option>
+      <gds-option value="robotics">Robotics</gds-option>
+      <gds-option value="quantum-computing">Quantum Computing</gds-option>
+      <gds-option value="biotechnology">Biotechnology</gds-option>
+      <gds-option value="genetic-engineering">Genetic Engineering</gds-option>
+      <gds-option value="holography">Holography</gds-option>
+      <gds-option value="fusion-power">Fusion Power</gds-option>
+      <gds-option value="antimatter">Antimatter</gds-option>
+      <gds-option value="dark-matter">Dark Matter</gds-option>
+      <gds-option value="neural-interfaces">Neural Interfaces</gds-option>
+      <gds-option value="exoskeletons">Exoskeletons</gds-option>
+      <gds-option value="space-elevators">Space Elevators</gds-option>
+      <gds-option value="terraforming">Terraforming</gds-option>
+      <gds-option value="nanobots">Nanobots</gds-option>
+      <gds-option value="smart-materials">Smart Materials</gds-option>
+      <gds-option value="augmented-reality">Augmented Reality</gds-option>
+      <gds-option value="virtual-reality">Virtual Reality</gds-option>
+      <gds-option value="brain-computer-interface"
+        >Brain-Computer Interface</gds-option
+      >
+      <gds-option value="cybersecurity">Cybersecurity</gds-option>
+      <gds-option value="biometrics">Biometrics</gds-option>
     </gds-dropdown>
   `,
 }