+++ libs/core/src/components/rich-text/rich-text.styles.ts
@@ -7,6 +7,7 @@ export const styles = css`
     --default-border: 1px solid
       color-mix(in srgb, currentColor, transparent 80%);
     letter-spacing: 0.0125rem;
+    display: block;
   }
 
   :host > :first-child {
@@ -55,12 +56,19 @@ export const styles = css`
   * lists
   * -----------------------------------------------------------
   */
-  ul,
-  ol,
+  ul {
+    padding-inline-start: 2ch;
+  }
+
+  ol {
+    margin-left: 3ch;
+    margin-bottom: 1lh;
+  }
+
   li {
-    margin: 0 0 0 1ch;
     padding: 0 0 0.3lh 0;
   }
+
   li:last-child {
     padding: 0;
   }
@@ -173,5 +181,10 @@ export const styles = css`
   details[open] summary {
     border-bottom: var(--gds-space-4xs) solid
       var(--gds-color-l3-border-tertiary);
+  }
+
+  a {
+    color: currentColor;
+    text-decoration-color: currentColor;
   }
 `