+++ libs/react/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.24.0
+
+### Minor Changes
+
+- 34bf172: **Dependencies:** Include React 19 in peer dependency constraint
+
+  This update makes it possible to use @sebgroup/green-react with React 19 without complaints from NPM.
+
+  Note, however, that the library is still not thoroughly tested with React 19, and all our unit tests still run against React 18. So proceed with caution.
+
 ## 3.23.1
 
 ### Patch Changes